Move APIClientUtils to test vone utils module
